+++ files/custom.py
@@ -1,0 +1,21 @@
+import hmac, hashlib
+
+
+"""
+Fichier de fonction utilisé par les fichiers du projet
+
+
+"""
+
+def customPRF512(key,A,B):
+    """
+    This function calculates the key expansion from the 256 bit PMK to the 512 bit PTK
+    """
+    blen = 64
+    i    = 0
+    R    = b''
+    while i<=((blen*8+159)/160):
+        hmacsha1 = hmac.new(key,A+str.encode(chr(0x00))+B+str.encode(chr(i)),hashlib.sha1)
+        i+=1
+        R = R+hmacsha1.digest()
+    return R[:blen]
